onConfrom can handle old format configuration
