+++ src/scripts/modules/wr-storage/adapters/destination.spec.def.js
@@ -41,83 +41,3 @@ export const cases = {
     }
   }
 };
-
-export const casesWithIncrement = {
-  disable: {
-    oldLocalState: {
-      destination: 'Test',
-      incremental: false
-    },
-    newLocalState: {
-      destination: 'Test',
-      mode: 'replace'
-    },
-    oldConfiguration: {
-      storage: {
-        input: {
-          tables: [
-            {
-              destination: 'Test'
-            }
-          ]
-        }
-      },
-      parameters: {
-        incremental: false
-      }
-    },
-    newConfiguration: {
-      storage: {
-        input: {
-          tables: [
-            {
-              destination: 'Test'
-            }
-          ]
-        }
-      },
-      parameters: {
-        mode: 'replace'
-      }
-    }
-  },
-
-  enable: {
-    oldLocalState: {
-      destination: 'Test',
-      incremental: true
-    },
-    newLocalState: {
-      destination: 'Test',
-      mode: 'update'
-    },
-    oldConfiguration: {
-      storage: {
-        input: {
-          tables: [
-            {
-              destination: 'Test'
-            }
-          ]
-        }
-      },
-      parameters: {
-        incremental: true
-      }
-    },
-    newConfiguration: {
-      storage: {
-        input: {
-          tables: [
-            {
-              destination: 'Test'
-            }
-          ]
-        }
-      },
-      parameters: {
-        mode: 'update'
-      }
-    }
-  }
-};
